ungit 1.5.1
Closes #47057.

+++ Formula/ungit.rb
@@ -3,8 +3,8 @@ require "language/node"
 class Ungit < Formula
   desc "The easiest way to use git. On any platform. Anywhere"
   homepage "https://github.com/FredrikNoren/ungit"
-  url "https://registry.npmjs.org/ungit/-/ungit-1.5.0.tgz"
-  sha256 "e41ec6128586f980140bb38393d4135710cb14d3ef3e4431720f1b3550ce6047"
+  url "https://registry.npmjs.org/ungit/-/ungit-1.5.1.tgz"
+  sha256 "8f045f6f606e3e89a8a053f86c9d403ba85df8d047e85a31a321fcce99cd9dee"
 
   bottle do
     cellar :any_skip_relocation
@@ -33,10 +33,12 @@ class Ungit < Formula
     sleep 5
     assert_match "ungit", Nokogiri::HTML(shell_output("curl -s 127.0.0.1:#{port}/")).at_css("title").text
   ensure
-    Process.kill("TERM", ppid)
-    # ensure that there are no spawned child processes left
-    child_p = shell_output("ps -o pid,ppid").scan(/^(\d+)\s+#{ppid}\s*$/).map { |p| p[0].to_i }
-    child_p.each { |pid| Process.kill("TERM", pid) }
-    Process.wait(ppid)
+    if ppid
+      Process.kill("TERM", ppid)
+      # ensure that there are no spawned child processes left
+      child_p = shell_output("ps -o pid,ppid").scan(/^(\d+)\s+#{ppid}\s*$/).map { |p| p[0].to_i }
+      child_p.each { |pid| Process.kill("TERM", pid) }
+      Process.wait(ppid)
+    end
   end
 end
